requirements.txt
Flask
torch
transformers
Pillow
tokenizers
